pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "isage-agentic"
version = "0.0.0.3"
description = "SAGE Agentic Framework - Agent framework, planning, tool selection, and workflow"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "IntelliStream Team", email = "shuhao_zhang@hust.edu.cn" }
]
keywords = [
    "agentic",
    "agent",
    "planning",
    "tool-selection",
    "workflow",
    "LLM",
    "AI",
]

dependencies = [
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
    "isage-libs>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.8.4",
    "mypy>=1.0.0",
]

llm = [
    "openai>=1.0.0",
    "anthropic>=0.20.0",
]

full = [
    "isage-agentic[dev,llm]",
]

[project.urls]
Homepage = "https://github.com/intellistream/sage-agentic"
Repository = "https://github.com/intellistream/sage-agentic"

[tool.setuptools]
# Root directory is the sage_agentic package
package-dir = {"sage_agentic" = "."}

# Explicitly list all packages (flat layout)
packages = [
    "sage_agentic",
    "sage_agentic.agents",
    "sage_agentic.agents.action",
    "sage_agentic.agents.action.tool_selection",
    "sage_agentic.agents.action.tool_selection.retriever",
    "sage_agentic.agents.bots",
    "sage_agentic.agents.planning",
    "sage_agentic.agents.planning.utils",
    "sage_agentic.agents.profile",
    "sage_agentic.agents.runtime",
    "sage_agentic.eval",
    "sage_agentic.interface",
    "sage_agentic.interface.protocols",
    "sage_agentic.interface.registries",
    "sage_agentic.interface.schemas",
    "sage_agentic.interfaces",
    "sage_agentic.reasoning",
    "sage_agentic.registry",
    "sage_agentic.workflow",
    "sage_agentic.workflow.generators",
    "sage_agentic.workflow.optimizers",
    "sage_agentic.workflows",
]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.json", "*.txt"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
